requirements.txt
Django>=3.1
channels>=2.4.0
asgiref==3.3.4 #https://github.com/django/asgiref/issues/278
channels_redis
django-registration-redux
django-extensions
django-oauth-toolkit
django-cors-headers
django-filter
django-taggit
django-probes
djangorestframework
drf-access-policy
psycopg2
ujson

django-channels-graphql-ws

django-guardian
django-avatar
python-slugify
graphene-django
unflatten
inflection
graphql-core<3,>=2.1.0
Markdown

# JWT
lok>=0.1.17

# GraphQL abstaction
balder[subscription]>=0.1.21

# Testing and Linting
ptvsd
pytest-django 
pytest-asyncio
pylint
pylint-django


pydantic
colorlog
namegenerator
whitenoise


# For Health Check Reasons
django-health-check
kombu
redis

pika
# Async Layer
aiormq
pyjwt
rich


#Deployment
omegaconf
gunicorn